adding script to convert xfel cxi files to simple hdf5
CXI files are hdf5 with some formatting. The main dataset is written in two parts
and a single file only seem to contain a single frame.

This script slurps up a whole bunch of cxi files (in a dir do *.cxi), and extracts the
two image parts into a numpy array. The numpy array is then offset by the largest absolute
min value to get the whole dataset above 0 - and then converted into a uint16 which works a
lot better for compression algorithms.